requirements.txt
requests>=2.28
pyyaml>=6.0

# Optional: install the provider(s) you plan to use
# anthropic>=0.30      # For Anthropic (Claude) provider
# openai>=1.0          # For OpenAI (GPT-4, DALL-E) provider
# No extra deps needed for Ollama, ComfyUI, or Stability providers (use requests)
